refactor: enum들 enums패키지로 분리
