+++ api/src/main/java/com/deliorder/api/enums/StoreStatus.java
@@ -1,4 +1,4 @@
-package com.deliorder.api.entity;
+package com.deliorder.api.enums;
 
 public enum StoreStatus {
     OPEN, CLOSED, BREAK
