+++ api/src/test/java/com/deliorder/api/service/StoreServiceTest.java
@@ -1,6 +1,9 @@
 package com.deliorder.api.service;
 
 import com.deliorder.api.entity.*;
+import com.deliorder.api.enums.DeliveryType;
+import com.deliorder.api.enums.DiscountType;
+import com.deliorder.api.enums.StoreStatus;
 import com.deliorder.api.repository.StoreRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
